Removed dependency
Service was only working when loaded after webViewer because of shared
ID. Removed id from constants (it is not a constant and shouldn't be
shared over this class) in the process.
Use of preferences accessible from different processes. (requires import
of script again, id can't be migrated)
Updated back toast to match long-press.
Moved return out of finally (never return in finally, it makes finally
useless)