Issue #5
It uses now only the new method to know if the index has changed.
But if it is not found and the previous one is, uses it instead, and
then removes it.
If both are found the new is used and the previous removed

So the issue is now fixed