Starts loading cached version before version check.
TODO implemented:
It starts showing the cached version if found.
Then it checks if there is a new version.
If so shows it.
(Note: due to the complexity of the loading/showing process, this
requires testing and may introduce bugs.)

Moved the loading indicator to the bottom right corner (because you may
want to interact with the page while it is loading)

Changed the noLauncher alert to shows that you can use the app without
it.